Rename FilterResults to Analysis, remove old analysis
